initial jquery prototype
all i need to do now is add all the greatness of the last release
to this version. i do lan on skipping all the xss stuff so i can
concentrate on the more important features. this means i will also
be writing a embeddable smartframe thingy to replace the old loader
code. next action is to get this back to ticking though.