support for nice artist data
artist data comes from gdocs during build time and
theres a form for adding data in the songtickerli
gui. im still thinking about how to do the history
browsing and a nice way for scrolling texts is also
not really done as i didnt like the way the previous
version handled that.